requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# Database packages for Amazon RDS
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
# Environment management
python-dotenv==1.0.0
# Additional utilities
pydantic-settings==2.1.0
